+++ analytics/redshift/redshift.go
@@ -1,5 +1,7 @@
 package redshift
 
+
+
 //CreateDatasets Create Datasets.
 func (redshift *Redshift) CreateDatasets(request interface{}) (resp interface{}, err error) {
 
@@ -13,20 +15,13 @@ func (redshift *Redshift) DeleteDatasets(request interface{}) (resp interface{},
 }
 
 
-type Describecluster struct{
-	clusterIdentifier string
-	marker string
-	maxRecords int
-	tagKeys []string
-	tagValues []string
-}
 
 //GetDatasets get Datasets.
 func (redshift *Redshift) GetDatasets(request interface{}) (resp interface{}, err error) {
 
 	var region string
 
-	var describecluster describecluster
+	var describecluster Describecluster
 
 	param := make(map[string]interface{})
 
@@ -57,7 +52,7 @@ func (redshift *Redshift) GetDatasets(request interface{}) (resp interface{}, er
 
 		case "TagValues":
 			tagValues, _ := value.([]string)
-			describecluster.TagValues = tagValues
+			describecluster.tagValues = tagValues
 		}
 	}
 
@@ -67,101 +62,11 @@ func (redshift *Redshift) GetDatasets(request interface{}) (resp interface{}, er
 	preparedefaultDescribeClusterspram(describeclusterspram)
 	prepareDescribeClusterspram(describeclusterspram,describecluster)
 
+	 response :=  make(map[string]interface{})
+
+	resp = PrepareSignaturequery(describeclusterspram,region,response)
+
  	return resp, err
-}
-
-
-func preparedefaultDescribeClusterspram(describeclusterspram map[string]string){
-
-	describeclusterspram["Action"]="DescribeClusters"
-  describeclusterspram["Version"]="2012-12-01"
-}
-
-func prepareDescribeClusterspram(describeclusterspram map[string]string,describecluster Describecluster){
-
-	describeclusterspram["Action"]="DescribeClusters"
-  describeclusterspram["Version"]="2012-12-01"
-
-	if describecluster.clusterIdentifier != ""{
-			describeclusterspram["ClusterIdentifier"] = describecluster.clusterIdentifier
-	}
-
-
-		if describecluster.marker != ""{
-				describeclusterspram["Marker"] = describecluster.marker
-	}
-
-	if describecluster.maxRecords != 0 {
-				describeclusterspram["MaxRecords"] = strconv.Itoa(describecluster.maxRecords)
-	}
-
-
-	if len(describecluster.tagKeys) != 0 {
-
-		for i := 0 ;i< len(describecluster.tagKeys); i++{
-
-			n := strconv.Itoa(i)
-
-			prefix = "TagKeys.TagKey." + n
-
-			describeclusterspram["prefix"] = describecluster.tagKeys[i]
-
-		}
-	}
-
-	if len(describecluster.tagValues) != 0 {
-
-		for i := 0 ; i< len(describecluster.tagValues); i++{
-
-			n := strconv.Itoa(i)
-
-			prefix = "TagValues.TagValue." + n
-
-			describeclusterspram["prefix"] = describecluster.tagValues[i]
-		}
-	}
-}
-
-
-func (ec2 *EC2) PrepareSignaturequery(describeclusterspram map[string]string, region string, response map[string]interface{}) error {
-
-	EC2Endpoint := "https://ec2." + regin + ".amazonaws.com"
-
-	req, err := http.NewRequest("GET", EC2Endpoint, nil)
-	if err != nil {
-		return err
-	}
-
-	// Add the params passed in to the query string
-	query := req.URL.Query()
-	for varName, varVal := range params {
-		query.Add(varName, varVal)
-	}
-	query.Add("Timestamp", timeNow().In(time.UTC).Format(time.RFC3339))
-
-	req.URL.RawQuery = query.Encode()
-
-
-	auth := map[string]string{"AccessKey": auth.Config.AWSAccessKeyID, "SecretKey": auth.Config.AWSSecretKey}
-
-	awsAuth.SignatureV2(req, auth)
-
-	r, err := http.DefaultClient.Do(req)
-
-	if err != nil {
-		return err
-	}
-	defer r.Body.Close()
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		return err
-	}
-
-	response["body"] = string(body)
-	response["status"] = r.StatusCode
-
-	return err
 }
 
 
@@ -174,6 +79,7 @@ func (redshift *Redshift) UpdateDatasets(request interface{}) (resp interface{},
 
 
 //ListDatasets  list Datasets.
-func (bigquery *Bigquery) ListDatasets(request interface{}) (resp interface{}, err error) {
+func (redshift *Redshift) ListDatasets(request interface{}) (resp interface{}, err error) {
+	return resp, err
 
 }
